+++ packages/google_maps_flutter/google_maps_flutter/ios/Classes/google_maps_flutter-umbrella.h
@@ -6,6 +6,7 @@
 #import <google_maps_flutter/FLTGoogleMapJSONConversions.h>
 #import <google_maps_flutter/FLTGoogleMapTileOverlayController.h>
 #import <google_maps_flutter/FLTGoogleMapsPlugin.h>
+#import <google_maps_flutter/GoogleMapHeatmapController.h>
 
 FOUNDATION_EXPORT double google_maps_flutterVersionNumber;
 FOUNDATION_EXPORT const unsigned char google_maps_flutterVersionString[];