+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.2.0
 
+* Adds support for heatmap layers.
 * Adds options for gesture handling and tilt controls on web.
 
 ## 2.1.8